Basic report investigating feasibility
Short report just looking at clinical remission and adverse events. More of a pilot analysis really to see what kind of numbers we get. Still need to be careful, especially with some of the outcomes which aren't covered by many papers which will likely lead to issues. I'll try to find what kind of ranges are acceptable or otherwise will base it off the Vedo paper.